==> R_markdowns/01_data_exploration.Rmd <==
---
title: "Data exploration"
subtitle: "From [Lizard guts: Naxos](https://github.com/katekathrynkat/lizard-guts-naxos)"
output:
  katereR::t_analysis
---

**Kate Culhane**, Dept. of Ecology, Evolution, and Marine Biology, UC Santa Barbara  
[kathrynculhane@ucsb.edu](kathrynculhane@ucsb.edu)
[<i class="fa fa-github"></i>](https://github.com/katekathrynkat/)
[<i class="fa fa-twitter"></i>](https://twitter.com/katekathrynkat/)

***

```{r setup, include=FALSE}
htmltools::tagList(rmarkdown::html_dependency_font_awesome())
knitr::opts_knit$set(root.dir='..')
knitr::opts_chunk$set(echo=TRUE, results='hold', warning=FALSE, message=FALSE,
                      fig.width=5, fig.height=3, rows.print=5)
```

<div class = "summary">

**SUMMARY**

Exploring the data

</div>

***

**DEPENDENCIES**

```{r dependencies, class.source='fold-show'}

##### PACKAGES #####

library(tidyverse) # data manipulation & visualization
library(kableExtra) # pretty tables
library(vegan) # community analyses
library(goeveg) # scree plot for NMDS

##### SOURCE CODE #####

source("R_scripts/aes_scales.R") # scales to use in ggplots
source("R_scripts/factor_levels.R") # functions to assign factor levels

##### DATA #####

# Habitat

# Abundance data
prey <- read_csv("output/00_prey.csv") %>% # sticky + pitfall samples
  fl_sample()
diet <- read_csv("output/00_diet_inverts.csv") %>% # diet samples (inverts)
  fl_sample()
diet_veg <- read_csv("output/00_diet_veg.csv") %>% # diet samples (veg)
  fl_sample()
inverts <- read_csv("output/00_inverts_all.csv") %>% # ALL invert samples
  fl_sample()

# Site by species matrices
prey_matrix <- read_csv("output/00_prey_matrix.csv") # sticky + pitfall samples
diet_matrix <- read_csv("output/00_diet_matrix.csv") # diet samples
inverts_matrix <- read_csv("output/00_inverts_matrix.csv") # ALL invert samples

```

## Abundance

<br>

**Number of invert individuals per order**

```{r}

inverts_kbl <- inverts %>% 
  group_by(order, sample_type) %>% 
  summarise(n = sum(count)) %>% 
  pivot_wider(names_from = sample_type, values_from = n) %>% 
  replace(., is.na(.), 0) %>% 
  mutate(trap_total = pitfall + sticky) %>% 
  relocate(diet, .after = last_col()) %>% 
  arrange(-trap_total)

kbl(inverts_kbl) %>% 
  kable_minimal(full_width = FALSE, position = "left") %>% 
  column_spec(1, bold = ifelse(inverts_kbl$diet > 10, "TRUE", "FALSE")) %>% 
  column_spec(5, bold = ifelse(inverts_kbl$diet > 10, "TRUE", "FALSE"),
                 color = ifelse(inverts_kbl$diet > 10, "red", "black"))
```

- Number of invert individuals summed across samples
- Orders with >10 individuals in diet samples highlighted in red

### Trapping samples

```{r fig.width=8, fig.height=8}

# Abundance by order
prey %>% 
  ggplot(aes(x = sample_type, y = count)) +
  geom_jitter(aes(color = sample_type),
              width = 0.1, size = 1.5) +
  scale_color_manual(values = scol_trap) +
  geom_boxplot(fill = NA, outlier.shape = NA) +
  facet_wrap(~ fct_reorder(order, count, .fun = sum, .desc = TRUE),
             scales = "free_y") +
  theme_classic() +
  theme(legend.position = "NA") +
  labs(x = "Trap type", y = "Abundance (indv/trap)")

```

- Number of invert individuals trapped per order
- Each point represents a trap (12 trap pairs/site, 9 sites)

### Diet samples

```{r fig.width=8, fig.height=5}

# Abundance by order
diet %>% 
  # Data wrangling
  group_by(order) %>% 
  mutate(n = length(count)) %>% 
  # Plot
  ggplot(aes(x = fct_reorder(order, count, .fun = sum, .desc = TRUE), y = count)) +
  geom_jitter(width = 0.1, size = 1.5) +
  geom_boxplot(fill = NA, outlier.shape = NA) +
  geom_text(aes(y = 10, label = paste("n =", n)),
            stat = "unique", angle = 90, size = 3, hjust = 0) +
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,11.5),
                     breaks = c(0,2,4,6,8,10)) +
  theme_classic() +
  theme(legend.position = "NA",
        axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1)) +
  labs(x = "Order", y = "Abundance (indv/lizard)")

```

- Number of invert individuals consumed per order
- Each point represents a stomach sample from a single lizard (148 lizards)
- Four outliers not plotted for hemiptera: 48, 22, 22, 17

## Dietary breadth

```{r}

diet_all <- full_join(diet, diet_veg, by = c("sample", "site", "sample_type",
                                             "date", "lizard_id",
                                             "order" = "veg", "count"))

```

## NMDS

### Prey (pitfall + sticky trap) data 

```{r cache=TRUE, results='hide'}

# Wrangle matrix
m_p <- as.matrix(prey_matrix[-c(1:4)])
rownames(m_p) <- prey_matrix$sample
m_p_hell <- decostand(m_p, 'hellinger') # Hellinger transformation
m_p_meta <- select(prey_matrix, sample, site, trap, sample_type) %>% # metadata
  fl_sample() %>% 
  addcol_useveg()

# Scree plot to check stress per number of dimensions
# dimcheckMDS(m_p_hell, distance = "bray", autotransform = FALSE, k = 10)
# Inflection point at k = 3

```

```{r cache=TRUE, results='hide', class.source='fold-show'}

# Create NMDS ordination
ord_p <- metaMDS(m_p_hell,
                 distance = 'bray', # use Bray-Curtis distances
                 autotransform = FALSE, # already manually transformed the matrix
                 k = 3, # number of dimensions
                 trymax = 1000)

# Stress plot
stressplot(ord_p)

```

#### By trap type

```{r fig.width=8, fig.height=6}

# Plot ordination
par(mar = c(4,4,1,1))
plot(ord_p, display = 'sites', type = 'n')
points(ord_p, display = 'sites', pch = 19,
       col = scol_trap[m_p_meta$sample_type])
ordiellipse(ord_p, groups = m_p_meta$sample_type,
            label = TRUE, col = scol_trap, lwd = 2)
text(ord_p, display = 'species')

```

- Each dot is a trap
- Number of dimensions = `ord_p$ndim`
- Stress = `round(ord_p$stress,3)`

#### By veg

```{r fig.width=8, fig.height=6}

# Plot ordination
par(mar = c(4,4,1,1))
plot(ord_p, display = 'species', type = 'n')
points(ord_p, display = 'sites', pch = 19,
       col = scol_veg[m_p_meta$veg])
ordiellipse(ord_p, groups = m_p_meta$veg,
            label = TRUE, col = scol_veg, lwd = 2)
text(ord_p, display = 'species')

```

#### By land use

```{r fig.width=8, fig.height=6}

# Plot ordination
par(mar = c(4,4,1,1))
plot(ord_p, display = 'species', type = 'n')
points(ord_p, display = 'sites', pch = 19,
       col = scol_use[m_p_meta$use])
ordiellipse(ord_p, groups = m_p_meta$use,
            label = TRUE, col = scol_use, lwd = 2)
text(ord_p, display = 'species')

```

### Lizard diet (stomach pumping) data

```{r cache=TRUE, results='hide'}

# Wrangle matrix
m_d <- as.matrix(diet_matrix[-c(1:4)])
rownames(m_d) <- diet_matrix$sample
m_d_hell <- decostand(m_d, 'hellinger') # Hellinger transformation
m_d_meta <- select(diet_matrix, sample, site, lizard_id) %>% # metadata
  addcol_useveg()

# Scree plot to check stress per number of dimensions
# dimcheckMDS(m_d_hell, distance = "bray", autotransform = FALSE, k = 10)

```

```{r cache=TRUE, results='hide', class.source='fold-show'}

# Create NMDS ordination
ord_d <- metaMDS(m_d_hell,
                 distance = 'bray', # use Bray-Curtis distances
                 autotransform = FALSE, # already manually transformed the matrix
                 k = 2, # number of dimensions
                 trymax = 1000)

# Stress plot
stressplot(ord_d)

```

#### By veg

```{r fig.width=8, fig.height=6}

# Plot ordination
par(mar = c(4,4,1,1))
plot(ord_d, display = 'sites', type = 'n')
points(ord_d, display = 'sites', pch = 19,
       col = scol_veg[m_d_meta$veg])
ordiellipse(ord_d, groups = m_d_meta$veg,
            label = TRUE, col = scol_veg, lwd = 2)
text(ord_d, display = 'species')

```

#### By land use

```{r fig.width=8, fig.height=6}

# Plot ordination
par(mar = c(4,4,1,1))
plot(ord_d, display = 'species', type = 'n')
points(ord_d, display = 'sites', pch = 19,
       col = scol_use[m_d_meta$use])
ordiellipse(ord_d, groups = m_d_meta$use,
            label = TRUE, col = scol_use, lwd = 2)
text(ord_d, display = 'species')

```

- Each dot is a trap
- Number of dimensions = `ord_d$ndim`
- Stress = `round(ord_d$stress,3)`


### Diet + trapping data

```{r cache=TRUE, results='hide'}

# Wrangle matrix
m_i <- as.matrix(inverts_matrix[-c(1:3)])
rownames(m_i) <- inverts_matrix$sample
m_i_hell <- decostand(m_i, 'hellinger') # Hellinger transformation
m_i_meta <- select(inverts_matrix, sample, site, sample_type) # metadata

# Scree plot to check stress per number of dimensions
dimcheckMDS(m_i_hell, distance = "bray", autotransform = FALSE, k = 10)

```

```{r cache=TRUE, results='hide', class.source='fold-show'}

# Create NMDS ordination
ord_i <- metaMDS(m_i_hell,
                 distance = 'bray', # use Bray-Curtis distances
                 autotransform = FALSE, # already manually transformed the matrix
                 k = 2, # number of dimensions
                 trymax = 1000)

# Stress plot
stressplot(ord_i)

```

```{r fig.width=8, fig.height=6}

# Plot ordination
par(mar = c(4,4,1,1))
plot(ord_i, display = 'species', type = 'n', xlim = c(-0.15, 0.15))

points(ord_i, display = 'sites', pch = 19,
       col = scol_sample[m_i_meta$sample_type])
ordiellipse(ord_i, groups = m_i_meta$sample_type,
            label = TRUE, col = scol_sample, lwd = 2)
text(ord_i, display = 'species')

```

- Each dot is a trap
- Number of dimensions = `ord_i$ndim`
- Stress = `round(ord_i$stress,3)`


<br>
<br>
<br>

***

<details><summary>**SESSION INFO**</summary>
```{r session-info}
devtools::session_info()
```
</details>
